Prevent premature shutdown of JVM when running a Server (#686)
Motivation:

Unless a user specified, Armeria uses the common boss and worker group
when building a Server. This can be a problem because all threads in the
common groups are daemon. If the user's main thread exits and there are
only the common group threads in the JVM, the JVM will choose to terminate
itself because all the remaining threads are daemon.

Modifications:

- Server now creates a dedicated single-thread boss EventLoopGroup for
  each ServerPort, and the thread is non-daemon.
  - Remove CommonPools.bossGroup
  - Remove ServerConfig.bossGroup and shutdownBossGroupOnStop
  - Server sets the thread name of boss groups dynamically to show which
    thread is handling which server port.
  - Server terminates the boss groups lastly so that the JVM does not
    terminate itself until the graceful shutdown is complete.
- Add Flags.numCommonBlockingTaskThreads() which replaces hard-coded 200.
- Add EventLoopGroups.newEventLoopGroup(numThreads, threadFactory)

Result:

- No more surprising JVM termination immediately after starting a Server up
- Ability to configure the number of common blocking task executor threads