Guarantee that RequestContext is accessible during converting a response by ResponseConverter. (#698)
Motivation:

An annotated method may return a future instance which runs over common ForkJoinPool, so we may get an error like ‘java.lang.IllegalStateException: RequestContext unavailable’  when accessing the context of a request in a RequestConverter. So we should guarantee that RequestContext is accessible during converting a response by ResponseConverter.

Modifications:

Call ‘RequestContext.push’ before converting a response by a ResponseConverter.

Result:

We are able to guarantee that RequestContext is accessible during converting a response by ResponseConverter.