Add a function that removes inside caches when the last subscriber starts to subscribe to StreamMessageDuplicator
Motivation:

StreamMessageDuplicator were keeping the signals before the `close()` is invoked due to it's own characteristic.
It doesn't, however, have to keep all the signals when it knows that there's no more subscription.
Therefore, it's needed to remove already published caches so that it prevents the internal cache growing largely.

Modifications:

- Add `duplicateStream(bool)` to set the last stream
- Replace internal list containing caches to a CircularQueue
- RetryingClient blocks the stream only configured amount of bytes to decide to retry

Result:

- Internal caches doesn't grow largely after the last subscriber starts to subscribes
- Close #622