Do not throw exception on unsupported HTTP method
Motivation:

Armeria's HttpMethod is an enum, and thus HttpMethod.valueOf() will
raise an exception when the given method name does not match any enum
values.

In contrast, Netty's HttpMethod is not an enum and thus supports
arbitrary method names.

As a result, when a user sends a request whose method is not supported
by Armeria's HttpMethod, the server will fail the request with an
unexpected IllegalArgumentException rather than sending '405 Method
Not Allowed' response.

Modifications:

- Add HttpMethod.isSupported(String)
- Add the validation of request method for HTTP/1 and 2

Result:

- Fixes #754
- Armeria responds with '405 Method Not Allowed' when a client sends a
  request with unsupported method.