Make EndpointSelector.select() aware of current request (#797)
Fixes #766 

Some load-balancing mechanism uses a property of the current request as its input, e.g. the current user's ID and the current request's certain parameter. However, it's not possible to implement such a mechanism with our endpoint API at the moment because `EndpointSelector.select()` does not accept any parameters. I propose to add `ClientRequestContext` to the parameter list of:

- `EndpointSelector.select()`
- `EndpointGroupRegistry.selectNode()`
- `Endpoint.resolve()`

so that a user can inject some information which could be used for load-balancing into `ClientRequestContext` attribute map.