Add EventLoopStreamMessage, which serializes stream operations throug an EventLoop (#855)
`EventLoopStreamMessage` is designed to have as few volatile reads as possible, going so far as to duplicate some fields to allow volatile reads in cold paths and non-volatile reads in hot paths (`state` and `isOpen` / `subscription` and `subscribed`). Notably, any `write` only has a single volatile read and store, while looping through objects has none at all.

The class is designed to be used when both the writer and reader thread are the same as the stream's. In such a case, there is no thread-switch overhead and we take advantage of it all being on the same thread to have almost no synchronization. For any other case, there will be heavy overhead from thread-switching - but Armeria is built on event loops, and it is extremely unlikely that the appropriate event loop cannot be used from writers and readers. Any case where it isn't possible probably isn't a performance sensitive application anyways.

This PR heavily refactors `DefaultStreamMessage` to allow logic-sharing between the two implementations, and any future ones (e.g., FixedStreamMessage #846 which I will rebase onto this one). One big design change is moving more logic out of `SubscriptionImpl` and into the publishers - by keeping `SubscriptionImpl` as a state-holder and method forwarder, it can be shared by any `StreamMessage` implementation.

An `AbstractStreamMessageTest` is also added to provide test cases for both implementations easily.