Treat RpcResponse.cause as log's cause if not explicitly set. (#862)
@anuraaga said:

In `GrpcService`, I call `responseContent` but not `endResponse`, and as a result `LoggingService` never sees any exception. While I could call `endResponse` as well, it was very unintuitive to me that setting a failure response wasn't enough to log the error, and feels repetitive to have to call twice with the same exception.