Close completed response before trying to abort its corresponding request
Motivation:
When HttpResponseWrapper is closed, it also aborts the corresponding request and the response is finally closed with a cause by HttpRequestSubscriber#failAndRespond via HttpRequestSubscriber#onError.
However, we should normally close it without a cause if the response has been completed. So we should give a chance to the HttpResponseWrapper to close the response normally before aborting the request.

Modifications:
- Call `request.abort()` after `delegate.close()` in HttpResponseDecoder.
- Close a request instead of aborting if the request is only handled by HttpServerHandler.

Result:
Closes #630