Add access logging feature to Server
Motivation:
Currently we cannot exactly write an access log for a server because our logging mechanism is based on a decorator and the decorator is only applied when a request is mapped to a service.
So it would be better to provide some access logging mechanism by a server itself, not a based on a decorator.

Modifications:
- Define log format types. Also provide a default log format such as NCSA common log format and NCSA combined log format.
- Add an interface to `ServerBuilder` to specify a log format and a log writer.
- Try to write an access log after a response has been sent.
- Add a site document.

Result:
Closes #817