Add a FixedStreamMessage which can be used to publish objects from a … (#846)
…stream fixed at construction time.

It is very common for an HttpRequest or HttpResponse to be constructed with all its objects known, e.g. using the respond family of methods. When all the objects are known, the synchronization overhead of DefaultStreamMessage is very wasteful and we can optimize this common use case.

This adds FixedStreamMessage, which serves objects straight from an indexed array. Since per reactive streams spec, only one thread can be used when notifying subscribers, we don't need synchronization when accessing the array and can iterate through these very quickly.

In the future, I would like to

Add extra-optimized FixedStreamMessage implementations for one or two object cases. Headers-only or Headers+content only are extremely common uses of FixedStreamMessage and it's worth removing the array for them.