Extract common logic for async benchmarks and allow running HttpServerBenchmark with H1C too (#898)
Immediately noticed that this level of DoS with an HTTP/1 client immediately causes `Too many open files`. I expected the files to be restricted to the number of event loops, but if not, then there needs to be a cap of some sort. Anyways, investigating that is for another PR wanted to get this cleanup / tweak in so others can try it out.

Disables request timeout and metrics too to better isolate performance.