Add request converter to annotated http service (#866)
Motivation:
When we are handling HTTP POST request, we usually convert its body to a Java object by some converting logic like JSON message converter. It is a general use-case for handling a request because a Java object is more convenient than a raw HTTP request itself. So it would be better if we have a request object mapping feature like Spring message converter.

Modifications:
- Add `@RequestConverter` annotation to specify a request converter of a method or a class. It is also repeatable.
- Add RequestConverterFunction interface to implement a request converter class.
- Add `@RequestObject` annotation to specify which parameter should be converted by a request converter.
- Add JacksonRequestConverterFunction and ByteArrayRequestConverterFunction as default request converters.