Change specifications of the response converter of the annotated service. (#907)
Motivation:
Recently we added `@RequestConverter` and `@ExceptionHandler` to annotated service. They have basically similar behavior such as:
- Iterating a converter (or handler) list when applying it.
- Using `accept()` method in order to check whether the converter (or handler) is able to consume the input.

`@Converter`, however, has its own behavior unlike the others. It maps a type of a result object to a converter so that it is only able to find a converter with the type. It is not customizable and is not useful in the real world.
So, it would be better to have them the same behavior and design as the `@RequestConverter` to make the response converter flexible.

Modifications:
- Rename `@Converter` to `@ResponseConverter` and `ResponseConverter` to `ResponseConverterFunction` for consistency.
- Rename method names of `*Function` interfaces so that a user can write a single class which inherits all interfaces.
- Change converter map to converter list.
- Add more `annotatedService()` methods to `ServerBuilder` and `AbstractVirtualHostBuilder`.
- Use default exception handlers and default request converters regardless of specifying them by a user.