Do not use Class.isAssignableFrom() if not applicable (#951)
Motivation:

We are using Class.isAssignableFrom() incorrectly in some places. For
example:

    expectedReturnType.isAssignableFrom(SomeType.class)

will return true even if 'expectedReturnType' is Object.class. We
obviously do not want our RequestConverterFunction implementations to
inject an Object parameter, but only when the parameter type matches in
a more meaningful way.

Similarly, I find this code:

    parameterInfo.type().isAssignableFrom(Optional.class)

which will match even when parameterInfo.type() is Object.class.

Modifictions:

- Use Class.isAssignableFrom() only when necessary

Result:

- More sensible type matching