Implement ServerListenerBuilder
Motivation:
Until now, ServerListener instance should be written by the user one by one. This approach is enough when there are only a few ServerListeners, but if lots of ServerListeners are required, this approach becomes not only cumbersome but also time-consuming and error-prone. Writing Armeria application that interacts with Zookeeper using Curator is a good example of this (#882) - you have to start and stop some Curator instances when the Server is started or stopping. Even worse, most of those routines are repetitive.

Modifications:
- Add `CallbackServerListener`: Now returns a concrete class which stores the callbacks as fields.
- Moved from Set operation to Add operation. Support `Iterable<Runnable>` API also.
- Add example to Javadoc.
- Refine `ServerListenerTest`.

Result:
- Easy to add ServerListeners