Make sure stub code handlers run on the event loop. (#1062)
Currently, methods called from gRPC stub code do not enforce event loop safety as they should. This has a reasonable chance of causing issues for the client in particular since it's common for stubs to be called from a server event loop, and especially true for unit tests.

Now, business logic is always run on the event loop. The few methods that are only for setting up state via experimental APIs before issuing any calls use synchronized instead to force a memory fence.

A test was removed which relied on the previous broken behavior of a write error being synchronously reported to the caller even when on a different thread.