Always call .join() on Server start futures in examples
Related: #1042

Motivation:

A user who is not used to asynchronous nature of our API may be confused
when a server fails to start up, because the failure is only notified
via the returned CompletableFuture.

Modifications:

Update all examples that calls Server.start() also call join() against
the returned future.

Result:

- Hopefully less confusing