Add Endpoint.ipAddr (#1089)
Motivation:

Currently, `Endpoint.host()` is used as a hostname or an IP address of
an endpoint. This does not work well for the `Endpoint`s whose IP
addresses are known already. For example, let's say we want to use
`DnsServiceEndpointGroup` with the following SRV records:

    _https._tcp.foo.com. 3600 IN SRV 10 100 443 one.foo.com.
    _https._tcp.foo.com. 3600 IN SRV 10 100 443 two.foo.com.
    _https._tcp.foo.com. 3600 IN SRV 10 100 443 three.foo.com.

Because the client will have to create a TLS connection with proper SNI
settings, the `Endpoint` should provide `foo.com` as the host name and
provide `<the IP address of {one,two,three}.foo.com>` as the IP address
of `foo.com`, which is not possible with our current API.

Modifications:

- Add a new property `ipAddr` to `Endpoint`
- Add `Endpoint.withIpAddr()`
- Handle the case where a host name in an `Endpoint` is an IP address
- Add test cases
- Compare authority, IP address and port only
  - Take `port` and `ipAddr` into account when comparing.
  - Do *not* take `weight` into account when comparing.
- Add host name validation using Guava InternetDomainName
- Fix incorrect uses of `Endpoint.of()` in some test cases
- Deprecate `Endpoint.of(host, port, weight)` in favor of
  `Endpoint.of(host, port).withWeight(weight)`

Result:

- Fixes #1087
- A user can create an `Endpoint` with an IP address.
- `Endpoint.equals()` makes more sense.