Add more flexibility in adding token-based authentication to HttpAuthServiceBuilder (#1132)
Currently if a user wants to use a bearer token on a header other than Authorization (perhaps they need to be able to assert two identities, a user and a server), it would require a lot of code duplication. This change allows users to specify a header name when adding token authorization. It also exposes a generic addTokenAuthorizer method which can be used for different token types too, there doesn't seem to be a great reason to keep the method private.