Add SAML configurator in order to easily enable SAML authentication on Armeria services
Motivation:
Nowdays a service doesn't need to provide a login service by itself because there are single sign-on specifications such as OAuth and SAML already.
So it would be better to support those specifications by Armeria so that a user can enable SSO on his or her services easily.

Modifications:
- Add saml module based on opensaml 3.x
- Add SamlServiceProvider class which helps a user enable SAML authentication for his or her services

Result:
Armeria provides SAML authentication for service providers.