Log request when response is failed and request wasn't logged. (#1244)
Currently, we allow setting different log levels for successful and failed responses. This is to allow the common case where successful requests / responses don't need to be logged (would be noise) but unsuccessful ones should for debugging. Unfortunately, the request is not logged for the latter right now, making it quite hard to debug. I think the default behavior should be to log the request in addition to the response when the response failed.

This change checks if request logging is enabled, and if not, logs request anyways if the response was unsuccessful. It admittedly feels a little weird, but I think it's less surprising for users. Open to any suggestions though.