Fix a StackOverflowError when stopping a starting server or starting a stopping server
Motivation:

When a user attempts to stop a server that's currently starting up or to
start a server that's currently shutting down, `Server` re-attempts to
stop/start the server after the current job (start/stop) is finished.

In a rare case, such a follow-up re-attempt is invoked even before the
state machine of the server gets a chance to enter `STOPPED` or `STARTED`
state, leading to infinite recursion.

Modifications:

- Use `handleAsync` rather than `handle` for potentially recursive
  operations, so that the start machine is given a chance to enter a new
  state.

Result:

`StackOverflowError` is gone.