Expose context-aware Executor for easier tests (#1307)
It is extremely common to use the `RequestContext`-aware eventloop as just an `Executor` for callbacks (e.g., `thenApplyAsync`). Unfortunately, when writing unit tests for such code, it can be cumbersome to mock `RequestContext` because there is no direct executing event loop (an event loop with an immediate executor usually deadlocks since callbacks get blocked by the previously running item in the loop). It slows down tests to create threads and is cumbersome to shut them down.

This exposes the exact same event loop as an `Executor`. It is purely to allow easier writing of tests where it them becomes possible to just do `when(ctx.contextAwareExecutor()).thenReturn(MoreExecutors.directExecutor())` for many cases where direct execution doesn't deadlock, allowing for faster, simpler tests.