Migrate docs client webapp to React / Typescript / Material UI (#1288)
This large PR migrates the docs client from a Javascript/jQuery/Bootstrap-based architecture to Typescript/React/MaterialUI. 

It adds a separate `docs-client` project which is a standard NodeJS / webpack based build, run using `gradle-node-plugin`. Currently it's a separate JAR, but if it makes more sense to add it into the `core` JAR pretty sure we can do that too.

I changed the URLs a little, `/method` -> `/methods`, `/namedType` -> `/structs` + `/enums`. Since struct and enum pages are quite different, it made sense to separate the URLs out. The old URls are still supported and redirect to the new ones.

The TypeScript configuration is as strict as possible, including whitespace / semantic linting based on AirBnb style + Prettier. It's a pretty common configuration nowadays.

Fixes #654