Add Copy and Clear buttons to docs client (#1321)
Please consider this PR as a feature request.

I added two buttons above the response section of the method page. They are the same level as Debug title, so should not comsume additional space.

<img width="1410" alt="screen shot 2018-08-06 at 15 24 08" src="https://user-images.githubusercontent.com/1726272/43700462-a688a45c-998d-11e8-8237-fb8b09c78966.png">

### Copy

Copies whole response text to the clipboard.

While using Armeria docs-client I often deal with big responses and it is often difficult to select and copy the whole response. This is where `Copy` button will be useful.

### Clear

Clears the response text.

Sometimes when I make the same request the second time it is hard to understand if I still see result of the last one or it is new one but the response is the same. In this case I reload the page to make the second request.

#### Compatibility

[execCommand](https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand) on the temp textarea is used to put text to the clipboard. It seems to be most common solution and supported by all popular browsers.

<img width="1136" alt="screen shot 2018-08-06 at 15 40 11" src="https://user-images.githubusercontent.com/1726272/43700918-5a99fd6e-998f-11e8-8989-5cd0ad859d23.png">