Set endOfStream for headers based on HTTP/2 frame EOS information. (#1325)
A simple proxy service in armeria is just essentially one line

```
(ctx, req) -> client.execute(req)
```

This should have the exact same semantics on the proxied client if it connects to the proxy or the backend, but it seems that because end-of-stream isn't set on received headers based on the HTTP/2 frame, this is not the case, and an empty data frame with EOS is inserted for trailers-only responses. We found this breaks a proxy to a gRPC backend where extra empty frames are not allowed.

The test fails before the code change and passes now.