Handle `%26` and `%3D` in a query string correctly
Motivation:

The following query string:

    a=1%26a=2

should be interpreted as a single query entry whose key is `a` and value
is `1&a=2`. However, `PathAndQuery` currently handles it as two query
entries: `a=1` and `a=2`.

It is because `PathAndQuery` decodes `%26` into `&` internally, making `%26`
indistinguishable from `&`.

Also, `%3D` in a query string should be similarly due to the case where
a query string key contains `=`: `a%3Db=1`

Modifications:

- Decode `%26` and `%3D` into a special marker character so that it is encoded
  back into `%26` and `%3D` in `PathAndQuery`
- Update test cases
  - ArmeriaCallFactory has been updated by @kojilin.

Result:

- A query parameter that contains an ampersand is interpreted correctly.