Use thread local when creating scopes off of the request context thread. (#1355)
While the `RequestContext` can be treated as the only thread-local for the request thread, this isn't the case for other threads, i.e., when making callbacks run in a threadpool context-aware. For such callbacks, we need to use a separate thread-local to prevent the multiple threads from clashing the context.