Report health for all endpoints, not just ones with unique authority. (#1375)
For an endpoint group like a DNS one, multiple endpoints have the same authority but different IP address. The current logic will only report health for one endpoint in such a case. Really, we want to make sure we report health for all endpoints, regardless of content, so this changes the logic to key on Endpoint instead of authority.

Only new tags are added, so it is basically backwards compatible.