Split RetryStrategy into RetryStrategy and RetryStrategyWithContent
Motivation:
When using `RetryingHttpClient`, we use `HttpResponseDuplicator` to duplicate the `HttpResponse` in order to determine to retry or not.
However, in most cases, we just need the response headers to decide and don't need the content.
If we see the content only when we really have to, we can minimize the usage of duplicator which reduces the performance. This approach works for `CircuitBreakerHttpClient` as well.

Modifications:
- Change the method of `RetryStrategy` to `shouldRetry(RequestContext ctx, Throwable cause)`
- Add `RetryStrategyWithContent`
- Add `CircuitBreakerClientBuiler`
- Change the method of `CircuitBreakerStrategy` to `shouldReportAsSuccess(RequestContext ctx, Throwable cause)`
- Add `CircuitBreakerStrategyWithContent`
- Add a benchmark to see the difference between using `Duplicator` and without it

Result:
- Better performance
```
Benchmark                 Mode  Cnt     Score      Error  Units
WithDuplicator.empty     thrpt    5  7187.756 ± 1695.714  ops/s
WithoutDuplicator.empty  thrpt    5  8844.105 ± 1395.332  ops/s
```
- Close #1408

Todo:
- Change `FixedStreamMessage` duplicable
- Don't let `RetryingHttpClient` use `HttpRequestDuplicator` when the content-length is zero