Don't rewrite the user's path if it is valid since it may be escaped. (#1417)
Users may have a specific reason to use special escaping in a client request (it may be the contract of the server such as [here](https://cloud.google.com/storage/docs/json_api/#encoding)). While the client should make sure request paths are valid, it should not overwrite whatever escaping the user may have used.

A side effect is double slashes are sent as is instead of rewriting them to single slashes. It's conceivable that a server requires double slashes for whatever reason and the client should support it (grudgingly).

Fixes #1416