Set RequestLog wall time in micros and use RequestLog timestamps for span timing (#1427)
JDK9+ supports returning wall time with microseconds precision. We may as well use it in logging instead of millis so users can take advantage of extra precision.

Notes:

I also went ahead and updated the Zipkin tracing to use the new values for spans. However I'm not exactly sure if this is good or bad - I think it means time spent in decorators can't be traced as the `RequestLog` is tied to the lifecycle of `StreamMessage`, not `Service`. On the flip side, maybe that is more natural behavior and it does make adding annotations like in #1423 much simpler.