Make One-way Thrift call responses sent early
Related: #1368 #1372
Motivation:

When a Thrift client sends a one-way call, it expects that the server
sends its response even before the call is actually handled.

Our current implementation sends a response after the call is handled,
which is not an expected behavior comparing to the official Thrift
implementation.

Modifications:

- Completed `RpcResponse` as soon as possible.
- Log at WARN level when a one-way call fails instead of failing
  `RpcResponse` because it's done already.

Result:

- Fixes #1368