Allow users to set additional trailers using ServiceRequestContext. (#1488)
It can be useful for RPC implementations, especially gRPC, to be able to specify additional trailers using `ServiceRequestContext`. 

This also updates `additionalResponseHeaders` manipulation to be thread-safe, which I don't think it was before. Compared to the previous implementation, `removeAdditional` always forces creation of headers but I think this is fine given usual usage of the method and makes sure it's thread-safe vs other add operations.

Fixes #1468