Marshall Status.cause between server and client for more structured error handling. (#1504)
Currently, when `verboseResponses` is enabled we serialize an exception's stack trace into the gRPC message, which is similar to the `message` of a `Throwable`. This allows the stack trace to be logged, e.g. by `LoggingClient` which is useful, but it could be even more useful if we provided better programmatic access to the stack trace.

This PR defines a proto which models the information found in an exception, including message, stack trace, and cause. Serializing this information allows the client to reconstruct almost all of the exception - the class itself is not used since the client may not have it or it may be difficult to instantiate it with reflection, so instead `StatusCauseException` provides information about the original class in its message and programmatically. Now, instead of the server's exception stack trace being logged as part of an exception's message, it is logged as part of an exception's cause, which gives more freedom to logging frameworks too.

The proto itself is serialized using gRPC's standard conventions for including proto in trailers, by using a key suffixed by `-bin` (handled by ProtoUtils) and a base64-encoded value. Non-armeria clients shouldn't have a huge trouble accessing the information in this proto if they need to.

Fixes #1471