Make Armeria(Client|Server)Configurator extend Spring's Ordered interface (#1535)
Motivation:
When specifying evaluation order of the configurators using Spring's `@Order` annotation, it's definitely inconvenient if you want to make a configurator to be called at the last. Because the default value of `@Order` annotation is `Integer.MAX_VALUE` which means the lowest precedence. So it would be more convenient if we change the default order of our configurators to `0` so that a user can easily make a configurator to be called at the last by specifying `@Order(1)` or `@Order(Ordered.LOWEST_PRECEDENCE)`.

Modifications:
- Make `Armeria(Client|Server)Configurator` extend `org.springframework.core.Ordered`.
  - Implement `getOrder()` as a default method which returns `0`.
- (Breaking change) Rename `ArmeriaClientConfigurator#customize` to `configure` for consistency.
- A user can specify multiple `ArmeriaClientConfigurator` beans, like `ArmeriaServerConfigurator`.

Result:
- Closes #1533