Make CORS policy to be applied to the specific paths
Motivation:
Currently, there is no way to apply a CORS policy to the specific paths. So it is hard for a user to configure different policies for different paths when the `Origin` domains are the same.
`@CorsDecorator` can handle this kind of situation because it can be specified on each methods in an annotated service, but using the annotation will be worked only if a user knows what origins are allowed at the compile time, which means that it's not suitable for configuring the origins dynamically. Also, it's only for the annotated services.
If the `CorsPolicy` has a `PathMapping` property, `CorsService` will be able to find a policy by using an `Origin` header value and the request path. Therefore, a user can apply CORS policies for different paths more convenient way.

Modifications:
- Added `boolean hasPathPatternOnly()` method to `PathMapping` interface, which describes whether the mapping has only path patterns as its condition.
- Added `pathMapping()` methods to `AbstractCorsPolicyBuilder` and `CorsServiceBuilder`.
  - Note that only the `PathMapping` that `PathMapping.hasPathPatternOnly()` returns `true` can be allowed.
- `CorsServiceBuilder` no longer keeps `ChainedCorsPolicyBuilder` list. It will be immediately built when its `and()` method is called. It's because the order of policies became important. The first configured path mapping must be evaluated first.
- Added `pathPatterns` property to `@CorsDecorator`.
- Updated `server-cors.rst`.

Result:
- A user can apply a CORS policy to the specific paths.
- Closes #1698.