Fix JSON parse on header in docs-client
Motivation:
When I try to use docs-client to simply test my API, I found that there
is an bug validation on header is not properly working.
I found that there is an error in onSubmit method. Armeria invokes
JSON.parse method to validate and parse user input header, but it does
not passed to conditional branch.

Modification:
I correct parsedHeader in checking branch onSubmit method.