Provide a way to get the duration that was taken to get a connection (#1692)
Motivation:
Currently, there's no way to get the duration for getting a connection. If you use HTTP/2,
getting the duration might be useless because, most of the time, you will share just one connection which was established beforehand. (So the duration is very short.) If you use HTTP/1.1, however,
you might need the value due to the nature of HTTP/1.1 which sends just one request at a time. (You need as many connections as you send requests concurrently.)

Modifications:
- Make `RequestLog` extends `AttributeMap` so that you can easily get the attributes from the log.
- Add `ClientConnectionTimings` and it's builder which has the timing information about a connection attempt. The timings are the start time and duration of:
  - `connectionAcquisition`
  - `dnsResolution`
  - `socketConnect`
  - `pendingAcquisition`
- Add the timing values to the client request metric.
- Miscellaneous
  - Calculate the length of request and response in `DefaultRequestLog` to use as the initial value of the `StringBuilder` for efficiency.
  - Deprecate `TextFormatter.epoch()` and `TextFormatter.appendEpoch()`, add
 `TextFormatter.epochMillis()`, `TextFormatter.appendEpochMillis()`, `TextFormatter.epochMicros()` and `TextFormatter.appendEpochMicros()`.

Result:
- Close #1678
- You can get the timing information about a connection attempt.